pyproject.toml
[tool.poetry]
name = "textsummarization"
version = "0.1.0"
description = "Analize text, create summary and sentiment"
authors = ["Alexander <as.frantsev@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.36.2"
gradio = "^4.11.0"
torch = "^2.2.0"
pydantic = "^2.5.3"
pretrainedmodels = "^0.7.4"
sentencepiece = "^0.1.99"
protobuf = "^4.25.1"
loguru = "^0.7.2"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.3"
aiohttp = "^3.9.3"
bcrypt = "^4.1.2"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
huggingface-hub = "^0.20.1"
flake8 = "^7.0.0"
black = "^24.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
